+pool . convert systems to singleton
added pool system and converted some managers to singleton for ease of use and to adhere to the proper designs. Also converted functionality that used instantiate/destroy to work with pools where necessary.

I deemed it necessary for highlight objects and not necessary for building production as there will not be that many buildings. even if there are it will only be created one at a time and not lead to issues.