InfoPanel Done.GridManager-> singleton.
it makes sense that gridmanager is singleton because anytime theres need for worldpos -> gridpos data needed itll be necessary to calculate the positions properly. also access to data on the grids is important for spawning units.

Next step is to spawn units and figuring out where to calll which data from. and building spawn unit logic for checking grid and returning the proper reply.