A*working, buggy attack range check
A* is a little bit buggy in its first step seems random at times.
the logic to step to the target positon is a bit weird but that might be due to position conversion from mouseposition to grid.
spawnmanager still be randomly calling stuff ondestroy somehow
other than that its smooth sailing .
Team system also implemented. for now its not being managed automatically but can be overridden on the inspector to test functionality.